Add feature flag ROUTER_GRACEFUL_SHUTDOWN_ENABLED

+++ router/src/main/resources/docker-entrypoint.sh
@@ -17,6 +17,9 @@
 
 set -e
 
+# Feature flag - Enable/disable the graceful shutdown
+export GRACEFUL_SHUTDOWN_ENABLED="${ROUTER_GRACEFUL_SHUTDOWN_ENABLED:-false}"
+
 echo "Configuring Choreo Connect Router"
 MG_ENVOY_YAML="$(envsubst < /home/wso2/envoy.yaml.template)"
 
@@ -37,7 +40,13 @@ ENVOY_PID=$!
 
 _term() {
     echo "Stopping Choreo Connect Router..."
-    kill -SIGTERM $ENVOY_PID
+
+    if [ "$GRACEFUL_SHUTDOWN_ENABLED" = "true" ]; then
+        echo "Graceful shutdown enabled. Sending SIGTERM to the envoy process..."
+        kill -SIGTERM $ENVOY_PID
+    else
+        echo "Graceful shutdown disabled. Not sending the SIGTERM to the envoy process..."
+    fi
     wait $ENVOY_PID
     echo "Choreo Connect Router stopped."
     exit 0
